[chaos] Rename staticpage component
