+++ src/core/static-page/index.css
